sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opengpl-sdk"
version = "0.1.0"
description = "The SDK for OpenGPL policy files — enforcement, authoring, and testing"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "click>=8.1",
    "presidio-analyzer>=2.2",
    "spacy>=3.7",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov", "httpx"]

[project.scripts]
opengpl = "opengpl.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/opengpl"]
